+++ lib/config.sh
@@ -194,13 +194,14 @@ function tredlyfile_parse() {
 
             # strip any whitespace
             local strippedValue=$(strip_whitespace "${value}")
-            local arrayKey
-            # extract data from the tredly definition
-            [[ $key =~ (.*)([[:digit:]]+)(.*)([[:digit:]]+)(.*) ]] && arrayKey="${BASH_REMATCH[2]}"
-            
+            local arrayKey=''
+            # extract relevant keys from the tredlyfile declaration
+            [[ ${key} =~ ^([a-zA-Z]+)([0-9]+) ]] && arrayKey="${BASH_REMATCH[2]}"
+
             # do different things based off certain commands
             case "${key}" in
                 url[1-999])
+                    
                     # add it to the array if it actually contained data
                     if [[ -n "${strippedValue}" ]]; then
                         _CONF_TREDLYFILE_URL[${arrayKey}]="${value}"
@@ -213,6 +214,7 @@ function tredlyfile_parse() {
                 url[1-999]Websocket)
                     # add it to the array
                     _CONF_TREDLYFILE_URLWEBSOCKET[${arrayKey}]="${value}"
+                    
                     ;;
                 url[1-999]MaxFileSize)
                     # validate it
@@ -235,6 +237,7 @@ function tredlyfile_parse() {
                             _CONF_TREDLYFILE_URLMAXFILESIZE[${arrayKey}]="1m"
                         fi
                     fi
+                    
                     ;;
                 url[1-999]Redirect[1-999])
                     # we can have multiple urlredirects per url so handle that with space separated string since space is encoded to %20 in urls
@@ -245,6 +248,7 @@ function tredlyfile_parse() {
                         # add it to the array
                         _CONF_TREDLYFILE_URLREDIRECT[${arrayKey}]="${value}"
                     fi
+                    
                     ;;
                 url[1-999]Redirect[1-999]Cert)
                     # if the value is blank then set the value to "null" so that the numbering within the string is correct
@@ -369,6 +373,21 @@ function tredlyfile_parse() {
     if [[ -z "${_CONF_TREDLYFILE[maxRam]}" ]]; then
         _CONF_TREDLYFILE[maxRam]="unlimited"
     fi
+    
+    # check if we have set "any" for any outgoing ports, if so then get rid of the rest of the ports as they are redundant
+    if array_contains_substring _CONF_TREDLYFILE_TCPOUT[@] '^any$'; then
+        _CONF_TREDLYFILE_TCPOUT=("any")
+    fi
+    if array_contains_substring _CONF_TREDLYFILE_UDPOUT[@] '^any$'; then
+        _CONF_TREDLYFILE_UDPOUT=("any")
+    fi
+    if array_contains_substring _CONF_TREDLYFILE_TCPIN[@] '^any$'; then
+        _CONF_TREDLYFILE_TCPIN=("any")
+    fi
+    if array_contains_substring _CONF_TREDLYFILE_UDPIN[@] '^any$'; then
+        _CONF_TREDLYFILE_UDPIN=("any")
+    fi
+    
     ## Skip the validation if need be
     if [[ (-z "${2}") || ("${2}" = false) ]]; then
         tredlyfile_validate
@@ -429,13 +448,31 @@ function tredlyfile_validate() {
         if [[ ${#_redirectUrls[@]} != ${#_redirectCerts[@]} ]]; then
             exit_with_error "url1: number of redirect urls does not match number of redirect certificates. If you are redirecting a HTTP URL, please include a blank definition."
         fi
-        
     done
     
-    # make sure that a mount point is specified if persistent storage is selected
-    if [[ -n "${_CONF_TREDLYFILE[persistentStorageUUID]}" ]] && ! array_contains_substring _CONF_TREDLYFILE_STARTUP[@] "^persistentMountPoint="; then
-        exit_with_error "'persistentStorageUUID' is specified but no mount point specified. Please specify a mount point in your Tredlyfile."
+    # ensure that HTTP or HTTPS ports are open if user specified a URL
+    if [[ ${#_CONF_TREDLYFILE_URL[@]} -gt 0 ]]; then
+        # loop over the urls
+        for _i in ${!_CONF_TREDLYFILE_URL[@]}; do
+            # check HTTP only urls - if urlcert is blank and port 80 isnt open
+            if [[ -z "${_CONF_TREDLYFILE_URLCERT[${_i}]}" ]] && \
+               ! array_contains_substring _CONF_TREDLYFILE_TCPIN[@] '^80$' && \
+               ! array_contains_substring _CONF_TREDLYFILE_TCPIN[@] '^any$'; then
+                exit_with_error "HTTP URL ${_CONF_TREDLYFILE_URL[${_i}]} specified but TCP IN port 80 is not set. Please set tcpInPort=80 in your Tredlyfile"
+            fi
+            # check HTTPS urls - if urlcert is not blank and port 443 isnt open
+            if [[ -n "${_CONF_TREDLYFILE_URLCERT[${_i}]}" ]] && \
+               ! array_contains_substring _CONF_TREDLYFILE_TCPIN[@] '^443$' && \
+               ! array_contains_substring _CONF_TREDLYFILE_TCPIN[@] '^any$'; then
+                exit_with_error "HTTPS URL ${_CONF_TREDLYFILE_URL[${_i}]} specified but TCP IN port 443 is not set. Please set tcpInPort=443 in your Tredlyfile"
+            fi
+        done
     fi
+    
+    # make sure that a mount point is specified if persistent storage is selected
+    #if [[ -n "${_CONF_TREDLYFILE[persistentStorageUUID]}" ]] && ! array_contains_substring _CONF_TREDLYFILE_STARTUP[@] "^persistentMountPoint="; then
+        #exit_with_error "'persistentStorageUUID' is specified but no mount point specified. Please specify a mount point in your Tredlyfile."
+    #fi
 
     ## Check that all the src paths in "fileFolderMapping" exist
     if [[ -n "${_CONF_TREDLYFILE[fileFolderMapping]}" ]]; then