+++ CHANGELOG.md
@@ -9,6 +9,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Implemented urlRedirect= for URLs
 - Implemented ability to change container resource limits/whitelist
   on the fly
+- Allow ports 80 and 443 (tcp) and 53 (udp) OUT by default
+  (no longer required to specify these in your Tredlyfile)
 
 ## 0.9.0 - 2016-04-21
 #### Added